+++ src/wininput.c
@@ -1515,6 +1515,7 @@ win_key_reset(void)
 
 #define dont_debug_virtual_key_codes
 #define dont_debug_key
+#define dont_debug_alt
 #define dont_debug_compose
 
 #ifdef debug_virtual_key_codes
@@ -1545,6 +1546,12 @@ vk_name(uint key)
 #define trace_key(tag)	printf(" <-%s\n", tag)
 #else
 #define trace_key(tag)	
+#endif
+
+#ifdef debug_alt
+#define trace_alt	printf
+#else
+#define trace_alt(...)	
 #endif
 
 // key names for user-definable functions
@@ -1857,10 +1864,13 @@ win_key_down(WPARAM wp, LPARAM lp)
   bool lalt = is_key_down(VK_LMENU);
   bool ralt = is_key_down(VK_RMENU);
   bool alt = lalt | ralt;
+  trace_alt("alt %d lalt %d ralt %d\n", alt, lalt, ralt);
   bool external_hotkey = false;
-  if (ralt && !scancode && cfg.external_hotkeys) {
+  // workaround moved down
+  if (0 && ralt && !scancode && cfg.external_hotkeys) {
     // Support external hot key injection by overriding disabled Alt+Fn
     // and fix buggy StrokeIt (#833).
+    trace_alt("ralt = false\n");
     ralt = false;
     if (cfg.external_hotkeys > 1)
       external_hotkey = true;
@@ -1869,8 +1879,18 @@ win_key_down(WPARAM wp, LPARAM lp)
   bool ctrl = lctrl | rctrl;
   bool ctrl_lalt_altgr = cfg.ctrl_alt_is_altgr & ctrl & lalt & !ralt;
   bool altgr = ralt | ctrl_lalt_altgr;
+  // workaround for StrokeIt
+  if (ralt && !scancode && cfg.external_hotkeys) {
+    // Support external hot key injection by overriding disabled Alt+Fn
+    // and fix buggy StrokeIt (#833).
+    trace_alt("ralt = false\n");
+    ralt = false;
+    if (cfg.external_hotkeys > 1)
+      external_hotkey = true;
+  }
   bool win = (is_key_down(VK_LWIN) && key != VK_LWIN)
           || (is_key_down(VK_RWIN) && key != VK_RWIN);
+  trace_alt("alt %d lalt %d ralt %d altgr %d\n", alt, lalt, ralt, altgr);
 
   mod_keys mods = shift * MDK_SHIFT
                 | alt * MDK_ALT
@@ -2468,6 +2488,7 @@ static struct {
     // https://web.archive.org/web/20120103012712/http://blogs.msdn.com/b/michkap/archive/2006/03/24/559169.aspx
     wchar wbuf[4];
     int wlen = ToUnicode(key, scancode, kbd, wbuf, lengthof(wbuf), 0);
+    trace_alt("layout %d alt %d altgr %d\n", wlen, alt, altgr);
     if (!wlen)     // Unassigned.
       return false;
     if (wlen < 0)  // Dead key.
@@ -2604,6 +2625,7 @@ static struct {
   }
 
   bool char_key(void) {
+    trace_alt("char_key alt %d -> %d\n", alt, lalt & !ctrl_lalt_altgr);
     alt = lalt & !ctrl_lalt_altgr;
 
     // Sync keyboard layout with our idea of AltGr.
@@ -2621,6 +2643,7 @@ static struct {
     if (ralt) {
       // Try with RightAlt/AltGr key treated as Alt.
       kbd[VK_CONTROL] = 0;
+      trace_alt("char_key ralt; alt = true\n");
       alt = true;
       layout();
       return true;
@@ -2632,6 +2655,7 @@ static struct {
     if (!altgr)
       return false;
 
+    trace_alt("altgr_key alt %d -> %d\n", alt, lalt & !ctrl_lalt_altgr);
     alt = lalt & !ctrl_lalt_altgr;
 
     // Sync keyboard layout with our idea of AltGr.
@@ -2713,6 +2737,7 @@ static struct {
     if (altgr) {
       // Try with AltGr treated as Alt.
       kbd[VK_CONTROL] = 0;
+      trace_alt("ctrl_key altgr alt = true\n");
       alt = true;
       return try_shifts();
     }
@@ -2896,6 +2921,9 @@ static struct {
       else if (VK_OEM_PLUS <= key && key <= VK_OEM_PERIOD)
         app_pad_code(key - VK_OEM_PLUS + '+');
     when VK_PACKET:
+      trace_alt("VK_PACKET\n");
+      if (altgr)
+        alt = lalt;
       if (!layout())
         return false;
     otherwise: